+++ framework/PageForm/PageForm.tsx
@@ -4,868 +4,33 @@ import {
   ActionGroup,
   Alert,
   Button,
-  Checkbox,
-  Flex,
-  FlexItem,
   Form,
-  FormGroup,
-  FormSection,
   Grid,
   gridItemSpanValueShape,
-  InputGroup,
-  Menu,
-  MenuItem,
-  PageSection,
-  Select,
-  SelectGroup,
-  SelectOption,
-  SelectOptionObject,
-  SelectVariant,
-  Slider,
-  Switch,
-  TextArea,
-  TextInput,
-  TextInputGroup,
-  TextInputGroupMain,
-  TextInputGroupUtilities,
   Tooltip,
 } from '@patternfly/react-core'
-import { CaretDownIcon, EyeIcon, EyeSlashIcon, SearchIcon } from '@patternfly/react-icons'
 import * as Ajv from 'ajv'
-import deepEqual from 'fast-deep-equal'
 import { JSONSchema6 } from 'json-schema'
-import {
-  Children,
-  CSSProperties,
-  Fragment,
-  isValidElement,
-  ReactNode,
-  useCallback,
-  useContext,
-  useState,
-} from 'react'
+import { CSSProperties, ReactNode, useContext, useState } from 'react'
 import {
   DeepPartial,
   ErrorOption,
   FieldPath,
   FieldValues,
   FormProvider,
-  SubmitHandler,
-  useController,
   useForm,
-  useFormContext,
-  UseFormReturn,
   useFormState,
 } from 'react-hook-form'
-import { PartialDeep } from 'type-fest'
 import { Scrollable } from '../components/Scrollable'
 import { useBreakpoint } from '../components/useBreakPoint'
-import { PageHeader, PageHeaderProps } from '../PageHeader'
 import { SettingsContext } from '../Settings'
-import { PageFormSelectOption } from './PageFormSelectOption'
-
-export type FormPageProps<T extends object> = PageHeaderProps & {
-  children?: ReactNode
-  defaultValues?: PartialDeep<T>
-  onSubmit: SubmitHandler<PartialDeep<T>>
-  schema?: unknown
-  isVertical?: boolean
-  onCancel?: () => void
-  submitText?: string
-  hideHeader?: boolean
-  noPadding?: boolean
-  form: UseFormReturn
-}
-export function FormPage<T extends object>(props: FormPageProps<T>) {
-  // const methods = useForm<PartialDeep<T>>({
-  //     defaultValues: props.defaultValues,
-  //     resolver: ajvResolver(props.schema, { strict: false }),
-  // })
-
-  const children = Children.toArray(props.children)
-
-  const inputs = children.filter((child) => {
-    if (!isValidElement(child)) return false
-    if (child.type === FormPageAlerts) return false
-    if (child.type === FormPageButtons) return false
-    return true
-  })
-
-  const buttons = children.find((child) => {
-    if (!isValidElement(child)) return false
-    if (child.type === FormPageButtons) return true
-    return false
-  })
-
-  const [settings] = useContext(SettingsContext)
-
-  return (
-    <>
-      {!props.hideHeader && <PageHeader {...props} />}
-      {/* <FormProvider {...methods}> */}
-      <Form
-        // eslint-disable-next-line @typescript-eslint/no-misused-promises
-        onSubmit={props.form.handleSubmit(props.onSubmit as SubmitHandler<FieldValues>)}
-        isHorizontal={props.isVertical ? false : settings.formLayout === 'horizontal'}
-        style={{
-          display: 'flex',
-          flexDirection: 'column',
-          flexGrow: 1,
-          overflow: 'hidden',
-        }}
-      >
-        <Scrollable style={{ height: '100%', flexGrow: 1 }}>
-          <PageSection
-            padding={{ default: props.noPadding ? 'noPadding' : 'padding' }}
-            isWidthLimited
-          >
-            <FormSection>{inputs}</FormSection>
-          </PageSection>
-        </Scrollable>
-        {buttons}
-      </Form>
-      {/* </FormProvider> */}
-    </>
-  )
-}
-
-export function FormPageAlerts() {
-  const { errors } = useFormState()
-  const isMd = useBreakpoint('md')
-  return (
-    <Fragment>
-      {errors && Object.keys(errors).length > 0 && (
-        <Alert
-          title="Please fix validation errors."
-          isInline
-          style={{ width: '100%', paddingLeft: isMd ? 190 : undefined }}
-          variant="danger"
-        >
-          {/* {process.env.NODE_ENV === 'development' && errors[Object.keys(errors)[0]].message} */}
-        </Alert>
-      )}
-    </Fragment>
-  )
-}
-
-export function FormPageButtons(props: {
-  submitText: string
-  cancelText: string
-  onCancel: () => void
-}) {
-  const { errors } = useFormState()
-
-  return (
-    <div>
-      <FormPageAlerts />
-      <PageSection
-        isFilled
-        style={{
-          display: 'flex',
-          flexDirection: 'column',
-          maxHeight: '100%',
-          borderTop: 'thin solid var(--pf-global--BorderColor--100)',
-        }}
-        variant="light"
-      >
-        <ActionGroup style={{ marginTop: 0 }}>
-          {errors && Object.keys(errors).length > 0 ? (
-            <Tooltip content={'Please fix validation errors'}>
-              <Button type="submit" isAriaDisabled>
-                {props.submitText}
-              </Button>
-            </Tooltip>
-          ) : (
-            <Button type="submit">{props.submitText}</Button>
-          )}
-
-          <Button type="button" variant="link" onClick={props.onCancel}>
-            {props.cancelText}
-          </Button>
-        </ActionGroup>
-      </PageSection>
-    </div>
-  )
-}
-
-export function FormInputCheckbox(props: {
-  label: string
-  name: string
-  helperText?: string
-  required?: boolean
-  description?: ReactNode
-  body?: ReactNode
-}) {
-  const { control } = useFormContext()
-  const { field } = useController({ control, name: props.name })
-  const id = props.name
-  return (
-    <Checkbox
-      label={props.label}
-      id={id}
-      aria-describedby={`${id}-helper`}
-      description={props.description}
-      body={field.value ? props.body : undefined}
-      {...field}
-      isChecked={!!field.value}
-    />
-  )
-}
-
-export function FormTextInput(props: {
-  id?: string
-  label: string
-  name: string
-  helperText?: string
-  required?: boolean
-  secret?: boolean
-  autoFocus?: boolean
-  placeholder?: string
-}) {
-  const {
-    register,
-    formState: { isSubmitting },
-  } = useFormContext()
-  const registration = register(props.name)
-  const { fieldState } = useController({ name: props.name })
-  const error = fieldState.error
-  const id = props.id ?? props.name
-  const [showSecret, setShowSecret] = useState(false)
-  return (
-    <FormGroup
-      id={`${id}-form-group`}
-      fieldId={id}
-      label={props.label}
-      helperText={props.helperText}
-      isRequired={props.required}
-      validated={error?.message ? 'error' : undefined}
-      helperTextInvalid={error?.message}
-    >
-      <InputGroup>
-        <TextInput
-          id={id}
-          type={props.secret && !showSecret ? 'password' : 'text'}
-          aria-describedby={`${id}-form-group`}
-          isRequired={props.required}
-          validated={error?.message ? 'error' : undefined}
-          autoFocus={props.autoFocus}
-          placeholder={props.placeholder}
-          {...registration}
-          onChange={(v, e) => {
-            void registration.onChange(e)
-          }}
-          isReadOnly={isSubmitting}
-        />
-        {props.secret && (
-          <Button
-            variant="control"
-            onClick={() => setShowSecret(!showSecret)}
-            aria-label="Options menu"
-            isDisabled={isSubmitting}
-          >
-            {showSecret ? <EyeIcon /> : <EyeSlashIcon />}
-          </Button>
-        )}
-      </InputGroup>
-    </FormGroup>
-  )
-}
-
-export function FormTextArea(props: {
-  id?: string
-  label: string
-  name: string
-  helperText?: string
-  required?: boolean
-  secret?: boolean
-  autoFocus?: boolean
-  placeholder?: string
-}) {
-  const {
-    register,
-    formState: { isSubmitting },
-  } = useFormContext()
-  const registration = register(props.name)
-  const { fieldState } = useController({ name: props.name })
-  const error = fieldState.error
-  const id = props.id ?? props.name
-  return (
-    <FormGroup
-      id={`${id}-form-group`}
-      fieldId={id}
-      label={props.label}
-      helperText={props.helperText}
-      isRequired={props.required}
-      validated={error?.message ? 'error' : undefined}
-      helperTextInvalid={error?.message}
-    >
-      <TextArea
-        id={id}
-        type={props.secret ? 'password' : 'text'}
-        aria-describedby={`${id}-form-group`}
-        isRequired={props.required}
-        validated={error?.message ? 'error' : undefined}
-        autoFocus={props.autoFocus}
-        placeholder={props.placeholder}
-        {...registration}
-        onChange={(v, e) => void registration.onChange(e)}
-        resizeOrientation="vertical"
-        isReadOnly={isSubmitting}
-        // innerRef={registration.ref}
-      />
-    </FormGroup>
-  )
-}
-
-export function FormNumberInput(props: {
-  id?: string
-  label: string
-  name: string
-  helperText?: string
-  required?: boolean
-  autoFocus?: boolean
-  min?: number
-  max?: number
-  valueLabel?: string
-}) {
-  const {
-    formState: { isSubmitting },
-  } = useFormContext()
-  const { field, fieldState } = useController({ name: props.name })
-  const error = fieldState.error
-  const id = props.id ?? props.name
-  const max = props.max ?? 100
-  const min = props.min ?? 1
-  const value = Number(field.value)
-  return (
-    <FormGroup
-      id={`${id}-form-group`}
-      fieldId={id}
-      label={props.label}
-      helperText={props.helperText}
-      isRequired={props.required}
-      validated={error?.message ? 'error' : undefined}
-      helperTextInvalid={error?.message}
-    >
-      <Flex alignItems={{ default: 'alignItemsFlexStart' }}>
-        <Flex
-          alignItems={{ default: 'alignItemsFlexStart' }}
-          spaceItems={{ default: 'spaceItemsSm' }}
-        >
-          <FlexItem style={{ paddingTop: 6, minWidth: 20, textAlign: 'right' }}>
-            {Math.floor((max - min) * value + min)}
-          </FlexItem>
-          {props.valueLabel && (
-            <FlexItem style={{ paddingTop: 6, minWidth: 40 }}>{props.valueLabel}</FlexItem>
-          )}
-        </Flex>
-        <FlexItem grow={{ default: 'grow' }}>
-          <Slider
-            id={id}
-            aria-describedby={`${id}-form-group`}
-            // isRequired={props.required}
-            // validated={error?.message ? 'error' : undefined}
-            autoFocus={props.autoFocus}
-            // {...registration}
-            value={(max - min) * value + min}
-            onChange={(v) => field.onChange((v - min) / (max - min))}
-            max={max}
-            min={min}
-            // innerRef={registration.ref}
-            isDisabled={isSubmitting}
-            showBoundaries={false}
-          />
-        </FlexItem>
-      </Flex>
-    </FormGroup>
-  )
-}
-
-export function FormBooleanInput(props: {
-  id?: string
-  label: string
-  name: string
-  helperText?: string
-  required?: boolean
-  autoFocus?: boolean
-}) {
-  const {
-    formState: { isSubmitting },
-  } = useFormContext()
-  const { field, fieldState } = useController({ name: props.name })
-  const error = fieldState.error
-  const id = props.id ?? props.name
-  return (
-    <FormGroup
-      id={`${id}-form-group`}
-      fieldId={id}
-      label={props.label}
-      helperText={props.helperText}
-      isRequired={props.required}
-      validated={error?.message ? 'error' : undefined}
-      helperTextInvalid={error?.message}
-    >
-      <Switch
-        id={id}
-        aria-describedby={`${id}-form-group`}
-        // isRequired={props.required}
-        // validated={error?.message ? 'error' : undefined}
-        autoFocus={props.autoFocus}
-        // placeholder={props.placeholder}
-        // {...registration}
-        isChecked={typeof field.value === 'boolean' ? field.value : false}
-        onChange={(e) => field.onChange(e)}
-        // innerRef={registration.ref}
-        isDisabled={isSubmitting}
-      />
-    </FormGroup>
-  )
-}
-
-export interface FormInputPopover {
-  title?: string
-  body?: string[]
-  link?: { label: string; url: string }
-}
-
-export interface FormSelectOption<T> {
-  group?: string
-  label: string
-  description?: string
-  value: T
-}
-
-export interface FormSelectProps<T> {
-  id?: string
-  name: string
-  label: string
-  popover?: FormInputPopover
-  placeholder?: string
-  help?: string
-  required?: boolean
-  footer?: {
-    label?: string
-    click?: () => void
-  }
-  create?: boolean
-  options: FormSelectOption<T>[]
-}
-
-export function FormSelect<T>(props: FormSelectProps<T>) {
-  const {
-    field,
-    fieldState: { error: fieldError },
-  } = useController({ name: props.name })
-  const [open, setOpen] = useState(false)
-  const id = props.id ?? props.name
-  const error = fieldError !== undefined && fieldError.message
-
-  const { options } = props
-  options.sort((l, r) => {
-    if ((l.group ?? '') < (r.group ?? '')) return -1
-    if ((l.group ?? '') > (r.group ?? '')) return 1
-    if (l.label < r.label) return -1
-    if (l.label > r.label) return 1
-    return 0
-  })
-  const selectedIndex = options.findIndex((option) => deepEqual(option.value, field.value))
-  const groups = options.reduce<
-    Record<
-      string,
-      {
-        group?: string
-        label: string
-        description?: string
-        value: T
-      }[]
-    >
-  >((groups, option) => {
-    const group = option.group ?? ''
-    let optionsArray = groups[group]
-    if (!optionsArray) {
-      optionsArray = []
-      groups[group] = optionsArray
-    }
-    optionsArray.push(option)
-    return groups
-  }, {})
-  const isGrouped =
-    Object.keys(groups).length > 1 ||
-    (Object.keys(groups).length === 1 && Object.keys(groups)[0] !== '')
-
-  const onSelect = useCallback(
-    (_event, value: string | SelectOptionObject) => {
-      if (typeof value !== 'string') {
-        value = value.toString()
-      }
-      const selectedIndex = Number(value)
-      const selected = options[selectedIndex].value
-      field.onChange(selected)
-      setOpen(false)
-    },
-    [field, options]
-  )
-  let index = 0
-  return (
-    <FormGroup
-      id={`${id}-form-group`}
-      fieldId={id}
-      label={props.label}
-      helperTextInvalid={error}
-      helperText={props.help}
-      isRequired={props.required}
-      validated={error ? 'error' : undefined}
-    >
-      <Select
-        id={id}
-        variant={SelectVariant.single}
-        aria-describedby={`${id}-helper`}
-        validated={error ? 'error' : undefined}
-        {...field}
-        isOpen={open}
-        onToggle={() => setOpen(!open)}
-        selections={selectedIndex === -1 ? '' : selectedIndex.toString()}
-        onSelect={onSelect}
-        isCreatable={props.create}
-        isInputFilterPersisted={props.create}
-        placeholderText={props.placeholder}
-        isGrouped={isGrouped}
-        ref={field.ref}
-        footer={
-          props.footer?.label && (
-            <Button
-              variant="link"
-              isInline
-              onClick={() => {
-                setOpen(false)
-                props.footer?.click?.()
-              }}
-            >
-              {props.footer.label}
-            </Button>
-          )
-        }
-        maxHeight={280}
-      >
-        {!isGrouped
-          ? options.map((option, index) => (
-              <SelectOption
-                key={index}
-                value={index.toString()}
-                label={option.label}
-                description={option.description}
-              >
-                {option.label}
-              </SelectOption>
-            ))
-          : Object.keys(groups).map((group) => (
-              <SelectGroup key={group} label={group}>
-                {groups[group].map((option) => (
-                  <SelectOption
-                    key={index}
-                    value={(index++).toString()}
-                    label={option.label}
-                    description={option.description}
-                  >
-                    {option.label}
-                  </SelectOption>
-                ))}
-              </SelectGroup>
-            ))}
-      </Select>
-    </FormGroup>
-  )
-}
-
-export function FormSelectInput(props: {
-  label: string
-  name: string
-  helperText?: string
-  required?: boolean
-  children?: ReactNode
-  footer?: ReactNode
-}) {
-  const { control } = useFormContext()
-  const {
-    field,
-    fieldState: { error },
-  } = useController({ control, name: props.name })
-
-  const menuItems = Children.toArray(props.children)
-    .filter((child) => isValidElement(child) && child.type === SelectOption)
-    .map((child) => {
-      if (isValidElement(child) && child.type === SelectOption) {
-        return (
-          <MenuItem key={(child.props as { value: string }).value}>
-            {(child.props as { children: ReactNode }).children}
-          </MenuItem>
-        )
-      }
-      return <></>
-    })
-
-  const [open, setOpen] = useState(false)
-  const id = props.name
-  let errorMessage: string | undefined
-  switch (error?.type) {
-    case 'required':
-      errorMessage = props.label + ' is required.'
-      break
-    default:
-      errorMessage = error?.type
-      break
-  }
-  return (
-    <FormGroup
-      fieldId={id}
-      label={props.label}
-      helperTextInvalid={errorMessage}
-      helperText={props.helperText}
-      isRequired={props.required}
-      validated={errorMessage ? 'error' : undefined}
-    >
-      <TextInputGroup>
-        <TextInputGroupMain {...field}></TextInputGroupMain>
-        <TextInputGroupUtilities>
-          {/* <Button variant="plain" onClick={() => {}} aria-label="Clear button and input">
-                        <TimesIcon />
-                    </Button> */}
-          <Button variant="plain" onClick={() => setOpen(!open)} aria-label="Options menu">
-            <CaretDownIcon />
-          </Button>
-        </TextInputGroupUtilities>
-        {open && (
-          <Menu style={{ position: 'absolute', right: 0, top: 36, width: '100%' }}>
-            {menuItems}
-          </Menu>
-        )}
-      </TextInputGroup>
-    </FormGroup>
-  )
-}
-
-export function FormTextSelect<T>(props: {
-  id?: string
-  label: string
-  name: string
-  helperText?: string
-  required?: boolean
-  secret?: boolean
-  autoFocus?: boolean
-  placeholder?: string
-  selectTitle?: string
-  selectValue?: (item: T) => string | number
-  selectOpen?: (callback: (item: T) => void, title: string) => void
-}) {
-  const {
-    register,
-    setValue,
-    formState: { isSubmitting },
-  } = useFormContext()
-  const registration = register(props.name)
-  const { fieldState } = useController({ name: props.name })
-  const error = fieldState.error
-  let id = props.id ?? props.name
-  id = id.split('.').join('-')
-  return (
-    <Fragment>
-      <FormGroup
-        id={`${id}-form-group`}
-        fieldId={id}
-        label={props.label}
-        helperText={props.helperText}
-        isRequired={props.required}
-        validated={error?.message ? 'error' : undefined}
-        helperTextInvalid={error?.message}
-      >
-        <InputGroup>
-          <TextInput
-            id={id}
-            type={props.secret ? 'password' : 'text'}
-            aria-describedby={`${id}-form-group`}
-            isRequired={props.required}
-            validated={error?.message ? 'error' : undefined}
-            autoFocus={props.autoFocus}
-            placeholder={props.placeholder}
-            {...registration}
-            onChange={(v, e) => {
-              void registration.onChange(e)
-            }}
-            // innerRef={registration.ref}
-            isReadOnly={isSubmitting}
-          />
-          <Button
-            variant="control"
-            onClick={() =>
-              props.selectOpen?.((item: T) => {
-                if (props.selectValue) {
-                  const value = props.selectValue(item)
-                  setValue(props.name, value, { shouldValidate: true })
-                }
-              }, props.selectTitle as string)
-            }
-            aria-label="Options menu"
-            isDisabled={isSubmitting}
-          >
-            <SearchIcon />
-          </Button>
-        </InputGroup>
-      </FormGroup>
-      {/* <SelectDialog
-                open={open}
-                setOpen={setOpen}
-                onClick={(organization: Organization) => {
-                    setValue(props.name, organization.name, { shouldValidate: true })
-                }}
-            /> */}
-    </Fragment>
-  )
-}
-
-export function FormSchema(props: { schema: JSONSchema6; base?: string }) {
-  const { schema } = props
-  const base = props.base ? props.base + '.' : ''
-
-  const p: ReactNode[] = []
-
-  for (const propertyName in schema.properties) {
-    const property = schema.properties[propertyName]
-
-    switch (property) {
-      case true:
-      case false:
-        continue
-    }
-
-    const title = typeof property.title === 'string' ? property.title : propertyName
-
-    let placeholder: string | undefined = (property as { placeholder?: string }).placeholder
-    placeholder = typeof placeholder === 'string' ? placeholder : undefined
-
-    const required = Array.isArray(schema.required) && schema.required.includes(propertyName)
-
-    switch (property.type) {
-      case 'string': {
-        switch ((property as { variant?: string }).variant) {
-          case 'select': {
-            if ('options' in property) {
-              const formSelectProps = property as unknown as FormSelectProps<unknown>
-              p.push(
-                <PageFormSelectOption
-                  key={base + propertyName}
-                  name={base + propertyName}
-                  label={title}
-                  placeholder={placeholder}
-                  isRequired={required}
-                  options={formSelectProps.options}
-                  footer={formSelectProps.footer}
-                />
-              )
-            } else {
-              p.push(
-                <FormTextSelect
-                  key={base + propertyName}
-                  name={base + propertyName}
-                  label={title}
-                  placeholder={placeholder}
-                  required={required}
-                  selectTitle={(property as { selectTitle?: string }).selectTitle}
-                  selectValue={
-                    (
-                      property as {
-                        selectValue?: (organization: unknown) => string | number
-                      }
-                    ).selectValue
-                  }
-                  selectOpen={
-                    (
-                      property as {
-                        selectOpen?: (callback: (item: unknown) => void, title: string) => void
-                      }
-                    ).selectOpen
-                  }
-                />
-              )
-            }
-            break
-          }
-          case 'textarea':
-            p.push(
-              <FormTextArea
-                key={base + propertyName}
-                name={base + propertyName}
-                label={title}
-                placeholder={placeholder}
-                required={required}
-              />
-            )
-            break
-          case 'secret':
-            p.push(
-              <FormTextInput
-                key={base + propertyName}
-                name={base + propertyName}
-                label={title}
-                placeholder={placeholder}
-                required={required}
-                secret
-              />
-            )
-            break
-          default:
-            p.push(
-              <FormTextInput
-                key={base + propertyName}
-                name={base + propertyName}
-                label={title}
-                placeholder={placeholder}
-                required={required}
-              />
-            )
-            break
-        }
-        break
-      }
-      case 'number': {
-        p.push(
-          <FormNumberInput
-            key={base + propertyName}
-            name={base + propertyName}
-            label={title}
-            required={required}
-            min={(property as { min?: number }).min}
-            max={(property as { max?: number }).max}
-            valueLabel={(property as { valueLabel?: string }).valueLabel}
-          />
-        )
-        break
-      }
-      case 'boolean': {
-        p.push(
-          <FormBooleanInput
-            key={base + propertyName}
-            name={base + propertyName}
-            label={title}
-            required={required}
-          />
-        )
-        break
-      }
-      case 'object': {
-        p.push(<FormSchema key={propertyName} schema={property} base={base + propertyName} />)
-        break
-      }
-    }
-  }
-
-  return <>{p}</>
-}
+import { FormSchema } from './PageFormFromSchema'
 
 export function PageForm<T extends object>(props: {
   schema?: JSONSchema6
   children?: ReactNode
   submitText: string
-  onSubmit: FormPageSubmitHandler<T>
+  onSubmit: PageFormSubmitHandler<T>
   cancelText: string
   onCancel?: () => void
   defaultValue?: DeepPartial<T>
@@ -967,7 +132,7 @@ export function PageForm<T extends object>(props: {
   )
 }
 
-export type FormPageSubmitHandler<T extends FieldValues> = (
+export type PageFormSubmitHandler<T extends FieldValues> = (
   data: T,
   setError: (error: string) => void,
   setFieldError: (fieldName: FieldPath<T>, error: ErrorOption) => void
