+++ framework/PageForm/PageFormInputPopover.tsx
@@ -1,0 +1,5 @@
+export interface PageFormInputPopover {
+  title?: string
+  body?: string[]
+  link?: { label: string; url: string }
+}
