+++ framework/index.ts
@@ -24,8 +24,8 @@ export * from './PageCells/LabelsCell'
 export * from './PageCells/TextCell'
 export * from './PageDataList'
 export * from './PageDialog'
+export * from './PageForm/Inputs/PageFormSelectOption'
 export * from './PageForm/PageForm'
-export * from './PageForm/PageFormSelectOption'
 export * from './PageFramework'
 export * from './PageHeader'
 export * from './PageLayout'
